+++ packages/benchmark-js/src/types.internal.ts
@@ -7,7 +7,7 @@ export type _Nanosecond = number & { __ns__: void };
 export interface _BenchmarkSettings {
     delay?: _Nanosecond;
     initCount?: number;
-    maxPrepareTime?: _Nanosecond;
+    maxPreparingTime?: _Nanosecond;
     maxTime?: _Nanosecond;
     minSamples?: number;
     minTime?: _Nanosecond;
@@ -50,3 +50,5 @@ export interface BenchmarkStats {
      */
     variance: number;
 }
+
+export type Hrtime = [number, number];