+++ packages/demo-test-nodejs/src/index.ts
@@ -1,3 +1,0 @@
-export {};
-
-console.log('This project is used for doing some small demo test.');