+++ packages/perf-js/src/string.ts
@@ -1,13 +1,18 @@
-import { Benchmark } from '@easy/benchmark-js';
+import { BenchmarkGroup } from '@easy/benchmark-js';
 
-/** Used to avoid some optimization */
-let res;
-res;
+const benchmark = new BenchmarkGroup();
 
-const benchmark = new Benchmark('String#indexOf', () => {
-    res = 'Hello World!'.indexOf('o') > -1;
+// let origin: string;
+// let target: string;
+const origin = 'Hello World!';
+const target = 'o';
+
+benchmark.add('String#indexOf-arg', () => {
+    origin.indexOf(target) > -1;
+});
+
+benchmark.add('String#indexOf', () => {
+    'Hello World!'.indexOf('o') > -1;
 });
 
 benchmark.run();
-
-console.log(benchmark.toString());